Move the vclock tests to the /test dir
