+++ test/test.rs
@@ -9,3 +9,4 @@ mod map;
 mod mvreg;
 mod orswot;
 mod pncounter;
+mod vclock;
